pyproject.toml
[tool.poetry]
name = "blog-gb"
version = "0.1.0"
description = ""
authors = ["Sego <grayg@mail.ru>"]
readme = "README.md"
packages = [{include = "blog_gb"}]

[tool.poetry.dependencies]
python = "^3.10.6"
flask = "==1.1.4"
flask-login = "==0.6.2"
flask-sqlalchemy = "==2.5.1"
python-dotenv = "==0.21.0"
psycopg2-binary = "==2.9.5"
flask-migrate = "==4.0.0"
flask-wtf = "==1.0.1"
email-validator = "==1.3.0"
flask-admin = "==1.6.0"
markupsafe = "==2.0.1"
flask-combo-jsonapi = "==1.1.0"
combojsonapi = "==1.1.2"

[tool.poetry.group.dev.dependencies]
black = "==21.12b0"
flake8 = "==6.0.0"
flake8-bugbear = "==22.12.6"
flake8-builtins = "==2.0.1"
pep8-naming = "==0.13.2"
flake8-commas = "==2.1.0"
flake8-variables-names = "==0.0.5"
flake8-unused-arguments = "==0.0.12"
isort = "==5.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
